pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "luthor"
version = "1.0.0"
description = "LLM Middleware"
authors = ["Lex Luthor"]
readme = "README.md"
include = [
    "git-commit.sha",
]

[tool.poetry.dependencies]
#############################################################################
# Centrally-managed dependencies
#
python = "3.12.3"
fastapi = "0.111.0"
pydantic = "2.7.1"
python-dotenv = "1.0.1"
pyyaml = "6.0.1"
typing-extensions = "4.12.2"
uvicorn = "0.30.1"

#############################################################################
# Add application-specific dependencies here
#

boto3-stubs = {extras = ["kms"], version = "1.34.128"}
httpx = "0.27.0"
pydantic-settings = "^2.0.3"
sqlalchemy = "2.0.30"  
pendulum = "3.0.0"
mypy-boto3-rds = "^1.34.116"
fastapi-filter = {extras = ["sqlalchemy"], version = "2.0.0"}
fastapi-debug-toolbar = "^0.6.3"


[tool.poetry.group.dev.dependencies]
#############################################################################
# Centrally-managed dev dependencies
#
coverage = "7.5.3"
mypy = "1.10.0"
poetry-plugin-bundle = "1.3.0"
pytest = "8.2.2"
pytest-mock = "3.14.0"
ruff = "0.4.9"
types-pyyaml = "6.0.12.20240311"
testcontainers = "4.4.0"

#############################################################################
# Add application-specific dev dependencies here
#
pytest-dotenv = "^0.5.2"
pytest-httpx = "0.30"
pytest-asyncio = "^0.23.7"
alembic = "1.13.1"
factory-boy = "^3.3.0"
freezegun = "^1.5.1"
pytest-alembic = "0.11.0"
pytest-xdist = "^3.6.1"
types-factory-boy = "^0.4.1"
sqlalchemy-utils = "^0.41.2"


[tool.ruff]
lint.ignore = ["E501"]
lint.select = ["E", "F", "I", "W"]
lint.fixable = ["I"]
line-length = 79

exclude = [".env", ".venv", "venv", "notebooks"]

[tool.coverage.paths]
source = ["src"]

[tool.coverage.run]
branch = true
relative_files = true
omit = [
  "tests/*",
]

[tool.coverage.report]
show_missing = true
fail_under = 80

[tool.mypy]
check_untyped_defs = true
disallow_any_unimported = true
disallow_untyped_defs = true
exclude = ["venv/", ".venv/"]
explicit_package_bases = true
ignore_missing_imports = true
mypy_path = "src"
no_implicit_optional = true
plugins = ["pydantic.mypy"]
python_version = "3.12"
show_error_codes = true
warn_return_any = true
warn_unused_ignores = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[tool.pyright]

reportCallIssue = false

[tool.pytest.ini_options]
env_files = [
    ".env.test"
]
